Always re-enqueue action requests generated by triggers
This change enforces that all action requests generated by triggers are
re-enqueued. This behaviour will ensure that triggered-actions behave in
a predictable way and prevents the situation that many high load
actions are run asynchronously without the worker waiting for them to
finish, resulting in unrestricted system load.
When the action request is enqueued it is now done with the originating
actor's session instead of a privileged session (which is how it worked
previously). There is no longer privilege escalation when executing an
action request generated by a triggered action.

This change also refactors test boostrap code to remove unneeded
abstractions and improves typing of a number of test files. Due to the
change to triggered action behaviour a number of tests that checked the
behaviour of triggers needed to be hoisted to a higher abstraction
level, as the previous test cases did not use the queue. This change
also revealed a bug where originator was not being passed when
a triggered action is enqueued.

Change-type: major
Signed-off-by: Lucian Buzzo <lucian.buzzo@gmail.com>