Revert change to behaviour in replaceCard when attachEvents:false
The behaviour of `Worker.replaceCard` is to always set `attachEvents` to false unless,
`options.attachEvents` is true AND there is no pre-existing contract.
The previous refactor incorrectly changed the behaviour to always ignore
`options.attachEvents`, which has now been fixed.
Long-term I really want to remove this method as it causes a bunch of
headaches in the core kernel and is a bit of a hack that appears to be
only used to bootstrap the stack.

Change-type: patch
Signed-off-by: Lucian Buzzo <lucian.buzzo@gmail.com>