Fix error when handling circular references in error objects
Previously, if an action handler threw an error containing a circular reference,
because errio doesn't handle circular references when serializing the
error, the kernel would attempt to insert the circular object and cause
a fatal error.
Switching from errio to serialize-error makes this code much safer, as
well as moving us away from errio, which has been deprecated (the GitHub
repo is actually archived). A confounder is that serialize-error@9.x is
an esm module, which makes integrating it into our dependency tree a bit
of a pain. For now I've opted to use serialize-error@8.x which is
a commonJS package and should play nicely. I would definitely consider
upgrading the tree to esm modules in the future.
See:
- https://gist.github.com/sindresorhus/a39789f98801d908bbc7ff3ecc99d99c

Change-type: patch
Signed-off-by: Lucian Buzzo <lucian.buzzo@gmail.com>