oauth: Fix bug where http request failures are not handled correctly
If a request to an external resource returns a failure code, the code
should be returned as-is. The axios behaviour is to throw on a status
code > 300, compared to the deprecated request package, which would just
return the status code in its callback params.
The rest of the code actually expects the >= 300 error codes and even has
logic to handle them sanely.

Change-type: patch
Signed-off-by: Lucian Buzzo <lucian.buzzo@gmail.com>