Process triggers in a separate task queue
Instead of waiting on triggers to run before resolving action execution,
queue up the task in an in-memory queue and process it asynchronously.

The results of evaluating triggers, transformer filters and subscriptions don't have an effect on the outcome of the action execution can be safely deferred, making the action execution more responsive.

Change-type: patch
Signed-off-by: Lucian Buzzo <lucian.buzzo@gmail.com>