Init worker with plugins
Update worker to be created with a list of plugins.
This means that all contracts provided by plugins
are inserted into the system when the worker starts.
Also allows us to add in the worker instance contract
update stream. A part of bringing the API codebase
into the worker.

Change-type: major
Signed-off-by: Josh Bowling <josh@balena.io>