Switch to polling strategy for cached contracts
Because updates to transformers, triggers and types are infrequent,
streaming these contracts generates a lot of overhead. Switching to
a slow polling strategy should reduce load on the DB.
This change also adds a `stop()` method that can safely shutdown the
worker, waiting for jobs to finish and stopping the polling.

Change-type: major
Signed-off-by: Lucian Buzzo <lucian.buzzo@gmail.com>