Use link from/to type as a heuristic when evaluating triggers
This change allows us to more quickly discard triggers that won't match
when evaluating link contract insertions.
The types on the trigger filter are compared to the types declared on
the link contract, and if they don't match we can safely exit the
trigger evaluation logic instead of having to query the kernel.

Change-type: patch
Signed-off-by: Lucian Buzzo <lucian.buzzo@gmail.com>