Fix bug where triggers are not recalculated when a relationship is added
Fixes a bug where some materialized fields won't be calculated correctly.
What happens is that because we now use relationships to create the triggered actions, used to process materialized fields, if you insert the relationship after the type then the triggered action won't be generated.  The triggered actions only get generated when a type contract is inserted.
We need to changes this so that the formulas will get parsed and the appropriate triggered actions generated when a new relationship is inserted.

Change-type: patch
Signed-off-by: Lucian Buzzo <lucian.buzzo@gmail.com>