Sync: Don't auto-add an origin field to synced contracts
The origin field can be used to debug synced data, as it is a reference to the original webhook event.
However this is largely redundant since we have request tracing via the log context id, and in the future will generate more complete graphs when executing action requests.
Removing this logic simplifies the code, and prevents situations where
contracts are needlessly updated.

Change-type: patch
Signed-off-by: Lucian Buzzo <lucian.buzzo@gmail.com>