Automatically subscribe users to direct message threads
Direct message threads are indicated by `dms: true` and `actors: [ 'user-1', 'user-2' ]`.
A triggered action detects new threads of this type being created and runs a dedicated action that will iterate over the actors participating in the thread and create a subscription on behalf of each of them.

Change-type: patch
Signed-off-by: Lucian Buzzo <lucian.buzzo@gmail.com>